+++ client/src/Components/Landing/WhatWeOffer/WhatWeOffer.jsx
